requirements.txt
requests
fastapi
python-dotenv
celery
redis
pymysql
uvicorn
sqlalchemy
